Link/Node: noop onClick/Hover if readonly
In  https://github.com/MrBlenny/react-flow-chart/issues/43 links and
nodes were clickable and deletable. This commit disables the onClick for
Nodes if config.readonly is set to true and disables onClick and Hover
for Links.

It also updates the OnLinkClick interface to point to IOnLinkClick.
Previously it pointed to IOnLinkMouseLeave which shares the same
interface but was likely a copy paste error.