Update port position when state changes
Update the position when the port is first mounted (this can happen
when the port "type" is changed e.g. from "bottom" to "right").

Also update the position when the number of ports on the port's side
changes.

Both of these updates are necessary to ensure that the links between
nodes start and end in the correct positions after an update to the
chart state.